Fix #4226 ca_name should not match fqdn
This commit provides the fix and updates the spec test for #4226
The fix changes the default behavior of the ca certificate common
name field.  When generating the CA certificate, the common name should
not exactly match the common name of the master's SSL certificate to
assist certificate revocation.

This commit changes the common name to be of the pattern:
CN=Puppet CA: $certname
e.g.
CN=Puppet CA: puppet.puppetlabs.lan

Signed-off-by: Jeff McCune <jeff@puppetlabs.com>