Add output when the list of faces is empty
When the list of faces is empty, the CLI user (me) is not presented with
any output running the "puppet faces list" command.  No error, no
information, nothing.

I tracked this down using rdebug.  What I don't understand and need to
find out is if faces.empty? should be true out of the box, or if we're
shipping some "default" faces or not.