WIP Improve metaprogramming error reporting
This change improves the error message for the end user of puppet.  The
root cause of the error is difficult for me to track down, but we can at
least be a little nicer about giving feedback.

The error message in the commit and this commit message should be
amended.