Use commands for plutil
The code had been executing plutil via backticks or %x{}, which is
hard to mock for systems that don't contain the plutil binary. To
remedy this, a commands statement was added so that a plutil method
is created off of an instance of the provider. This makes it easier
to mock for testing.

Stub plutil command

Spec tests were failing on Linux systems that didn't contain the
plutil binary. To remedy this, plutil needs to be stubbed.

WIP - Tightening up mocks in spec tests