(#11868) Use Windows Installer interfaces to manage MSI packages
Previously, we were managing MSI packages using the msiexec.exe command
line utility. However, it did not provide a way to query the system to
determine what packages are already installed. Also to uninstall a
package, puppet had to specify either the original source or the product
code GUID, but puppet didn't have access to the latter.

As a result, puppet recorded the product name and source of each product
it installed in a state yaml file.  During uninstall puppet would
retrieve the source parameter for the package and pass that along to
msiexec.exe.

However, this lead to several problems. Puppet only knew about products
it installed, so it couldn't remove products it didn't install. Also if
someone removed a product via Add/Remove Programs, puppet wouldn't know
about it, and report that the product was still installed. Worse, if
puppet's var directory was deleted, then it would lose all of its
product state information.

This commit changes the msi package provider to use the Windows
Installer Automation interfaces to install, query, and uninstall
products. In doing so, it will correctly report on products it didn't
install, including puppet itself (#13444). It also can detect when a
product is removed, and re-install it.

Another benefit is that the Windows Installer supports UNC and URL
sources, so the following are now possible:

  package { 'splunk':
    source => '\\server\share\splunk-version.msi'
  }

  package { 'msys':
    source => 'http://host:port/msys.msi'
  }

Although puppet can now detect what versions of software are installed,
it cannot yet implement the `versionable` feature, as that would require
examining registered patches for a given product, and seeing if one of
the patches satifies the version requirement.