Use autoloader in application / command line methods
There were a few places where we were loading things
directly from the LOAD_PATH, rather than going through
the autoloader; now that we are considering having
the modulepath lib dirs eligible for loading from,
it is important that we handle all of this in a single,
central spot in the code.  Thus this commit changes
the command_line/application/face_collection classes
to use the autoloader rather than trying to manage
things themselves.