Remove confdir and vardir from run_mode
Prior to this commit, the vardir and confdir settings
were tied to run_mode.  This caused a chicken-and-egg
problem where you could not initialize the confdir
and vardir until you'd loaded the application/subcommand,
because that is the first point in time where you'd know
the run_mode.  However, without knowing the confdir/vardir
ahead of time, it's impossible to know where all you might
need to look to try to find the application/subcommand for
loading.

This commit decouples those settings from run_mode.  The
default values for them are now set solely based on whether
the user is root or not.  Individual applications may
override these defaults a bit later in the lifecycle
(e.g., 'master' does this).