Add support for gem dependency list in Modulefile
This commit adds support for a "gems" line in a
puppet modulefile.  If the line exists, then the
module tool will use a puppet package provider
to attempt to install all of the gems that are
prerequisites of the module.

This commit is functional but is really only
for illustration purposes; it has no tests and
is lacking a lot of important error handling...
it's basically just a proof-of-concept.