Create base64_decode_string
Previously, the code for Base64 decoding a hex string into
a binary string was provided inline within the body of the
provider methods that needed this behavior.  This meant that
code was duplicated, but, more importantly, to test the
behaviour you would need to mock out the Base64.decode64
method, which is not the best case in the world.

To remedy this, this commit creates a new method, base64_decode_string,
and wraps the necessary behaviour within that method.  This should
reduce code duplication and also make it easier to test.