Update spec tests
This commit updates the spec tests for the directoryservice
user provider according to recent commits that introduced
helper methods for common behavior.  These tests should be
tighter around behavior that was difficult to test (testing
StringIO objects), as well as adding tests to catch edge-
cases (10.6-style users that exist in 10.8).