(#7911) Fixed Libuser.getconf, updated unit specs, removed conflicting expiry fix
This commit change the Puppet::Util::Libuser module to look for libuser.conf in its new /etc/puppet/provider/libuser.conf location.  It also updates Puppet::Features::Libuser to only return true if the config file as well as the binaries are found.

Updated the useradd and groupadd specs to expect :execute when called from create to now have 2 parametes.  A command and an options hash.

During my work on this feature I added an exipry getter since the builtin getter didn't retrieve the correct values.  Before this patch was merged another commit came along that also set the proper getter. So I removed mine as it caused a conflict.