(#16451) Look for systemctl in both /bin and /usr/bin
Archlinux has moved the systemctl binary from `/bin/systemctl` to
`/usr/bin/systemctl`. As a result the `systemd` service provider does
not work anymore because the necessary binaries cannot be found.

Check if systemctl is installed under /usr/bin/systemctl and fall back
to /bin/systemctl if not found. This way we support both versions and
the provider will still be considered unsuitable if neither
/usr/bin/systemctl nor /bin/systemctl is present.