(#16451) Use PATH to find systemctl for systemd service provider
Without this patch Puppet has a difficult time locating the systemctl
executable in recent versions of Archlinux.  This problem is caused by
the executable moving from /bin/systemctl to /usr/bin/systemctl.

This patch addresses the problem by eliminating the fully qualified path
and instead relying on the PATH environment variable to locate the
command.