(maint) Enable http(s) Git URLs in the Gemfile
Without this patch it is not possible to use Bundler to install
dependencies from Git repositories using the HTTP or HTTPS protocols.
This is a problem because we want to configure our CI system to proxy
requests through proxy.puppetlabs.lan:3128.

Using HTTP will allow us to cache results, saving network bandwidth and
reducing reliance on external service availability.