(maint) Fix failing WriteOnlyYaml acceptance test
Without this patch the WriteOnlyYaml acceptance test is failing in
master.  The root cause of this failure is the change in behavior in
30fe09e - (#20284) Output one item per line for arrays in console output

The change in behavior causes the default renderer to print a more human
readable verticle list when the output data from Faces is an array.
This change caused the assumptions made in the WriteOnlyYaml test to no
longer be valid.

The acceptance tests appears to assume JSON output from the `puppet
node` command in order to validate the presence of the expected cached
node data.  This patch changes the test to explicitly render the output
of the command using json to ensure the command produces the output
format the test expects.